+++ src/app/modules/admin/catalogs/products/list/products-list.component.ts
@@ -13,6 +13,8 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
+import Swal from "sweetalert2";
+
 
 @Component({
   selector: 'app-products-list',
@@ -90,11 +92,16 @@ export class ProductsListComponent implements OnInit {
     this.router.navigate([`/catalogs/products/${ProductId}`]); // Ajustado a la ruta correcta
   }
 
-  deleteProduct(ProductId: number): void {
-    this.productsService.deleteProduct(ProductId).subscribe(() => {
-      this.getProducts();
-      this.snackBar.open('Producto eliminado correctamente', 'Cerrar', { duration: 3000 });
-    });
+  async deleteProduct(ProductId: number){
+    const confirmed = await this.showConfirmation();
+    if (confirmed) {
+      this.productsService.deleteProduct(ProductId).subscribe(() => {
+        this.getProducts();
+        this.snackBar.open("Proyecto eliminado correctamente", "Cerrar", {
+          duration: 3000,
+        });
+      });
+    }
   }
 
   obtenerPermisos(): void {
@@ -136,4 +143,17 @@ export class ProductsListComponent implements OnInit {
   tienePermiso(codigo: string): boolean {
     return this.permisosDisponibles.includes(codigo);
   }
+  showConfirmation(): Promise<boolean> {
+        return Swal.fire({
+          title: "Seguro que desea eliminar",
+          text: "Esta accion no se puede revertir",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonText: "Eliminar",
+          cancelButtonText: "Cancelar",
+          reverseButtons: true,
+        }).then((result) => {
+          return result.isConfirmed;
+        });
+      }
 }
